GLOBAL: Integrate TI-NSPIRE port (#90)
* nspire: Add initial

* NSPIRE: Remove -Werror and -fanalyzer

* .gitignore: Modify to exclude .tns files

* Add Tyler's fixes for latest SDK

* Remove yml

* nspire: Render zombie further than alias model

* Re-add github actions

* NSPIRE: port restructure from upstream vril

* Update YAML to add TNS build

* NSPIRE: Fix blood printing

* Support compiling with -Werror

* Remove fixed div for non nspire

* Revise makefile build name and actions

* Update README.md

* Add large texture gating

* Update README.md

* Revert condebug hack

* COMMON: Revise NSPIRE specific blood

* Migrate more NSPIRE specifics to common with stubs

* NSPIRE: Limit ifdef in nzportable_def

* TNS: Add Vril features to clean up Nspire-specific common changes

---------

+++ source/cl_slist.c
@@ -45,7 +45,7 @@ char *CreateSpaces(int amount)
 	static char spaces[1024];
 	int size;
 
-	size = bound(1, amount, sizeof(spaces) - 1);
+	size = bound(1, amount, (int)sizeof(spaces) - 1);
 	memset(spaces, ' ', size);
 	spaces[size] = 0;
 
@@ -180,7 +180,7 @@ void SList_Load (void)
 		}
 		else
 		{
-			if (len + 1 < sizeof(line))
+			if (len + 1 < (int)sizeof(line))
 				line[len++] = c;
 		}
 	}
